feat: break button board tasks more reliably
Previously task has been broken so that the switch state is
ensured to be different than last time, however if the
switch has been flipped since, the task could be
immediately fixed.  Check instead the current position of
the switch and break the task so it needs to be flipped.